+++ src/Either.ts
@@ -361,6 +361,12 @@ export const getOrElseW = <E, B>(onLeft: (e: E) => B) => <A>(ma: Either<E, A>): 
  */
 export const getOrElse: <E, A>(onLeft: (e: E) => A) => (ma: Either<E, A>) => A = getOrElseW
 
+/**
+ * @category destructors
+ * @since 2.10.0
+ */
+export const toUnion = <E, A>(fa: Either<E, A>): E | A => pipe(fa, fold<E, A, E | A>(identity, identity))
+
 // -------------------------------------------------------------------------------------
 // combinators
 // -------------------------------------------------------------------------------------